requirements.txt
streamlit
openai
pandas
pyyaml
python-dotenv
